update and rename image_dir format
